pyproject.toml
[tool.poetry]
name = "template-this"
version = "0.1.0"
description = ""
authors = ["NyxAether <contact.nyxhemera@gmail.com>"]
readme = "README.md"
packages = [{include = "template_this"}]

[tool.poetry.dependencies]
python = "^3.11"
black = "^24.2.0"
mypy = "^1.8.0"
pylint = "^3.0.3"
flake8 = "^7.0.0"
pydantic = "^2.6.1"
click = "^8.1.7"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.12"
isort = "^5.13.2"
autoflake = "^2.2.1"
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tt = "template_this.cli:main"